Migrate perform required password reset endpoint
