+++ server/service/handler.go
@@ -24,16 +24,15 @@ import (
 
 // FleetEndpoints is a collection of RPC endpoints implemented by the Fleet API.
 type FleetEndpoints struct {
-	Login                        endpoint.Endpoint
-	Logout                       endpoint.Endpoint
-	ForgotPassword               endpoint.Endpoint
-	ResetPassword                endpoint.Endpoint
-	CreateUserWithInvite         endpoint.Endpoint
-	PerformRequiredPasswordReset endpoint.Endpoint
-	VerifyInvite                 endpoint.Endpoint
-	InitiateSSO                  endpoint.Endpoint
-	CallbackSSO                  endpoint.Endpoint
-	SSOSettings                  endpoint.Endpoint
+	Login                endpoint.Endpoint
+	Logout               endpoint.Endpoint
+	ForgotPassword       endpoint.Endpoint
+	ResetPassword        endpoint.Endpoint
+	CreateUserWithInvite endpoint.Endpoint
+	VerifyInvite         endpoint.Endpoint
+	InitiateSSO          endpoint.Endpoint
+	CallbackSSO          endpoint.Endpoint
+	SSOSettings          endpoint.Endpoint
 }
 
 // MakeFleetServerEndpoints creates the Fleet API endpoints.
@@ -56,24 +55,19 @@ func MakeFleetServerEndpoints(svc fleet.Service, urlPrefix string, limitStore th
 		InitiateSSO:          logged(makeInitiateSSOEndpoint(svc)),
 		CallbackSSO:          logged(makeCallbackSSOEndpoint(svc, urlPrefix)),
 		SSOSettings:          logged(makeSSOSettingsEndpoint(svc)),
-
-		// PerformRequiredPasswordReset needs only to authenticate the
-		// logged in user
-		PerformRequiredPasswordReset: logged(canPerformPasswordReset(makePerformRequiredPasswordResetEndpoint(svc))),
 	}
 }
 
 type fleetHandlers struct {
-	Login                        http.Handler
-	Logout                       http.Handler
-	ForgotPassword               http.Handler
-	ResetPassword                http.Handler
-	CreateUserWithInvite         http.Handler
-	PerformRequiredPasswordReset http.Handler
-	VerifyInvite                 http.Handler
-	InitiateSSO                  http.Handler
-	CallbackSSO                  http.Handler
-	SettingsSSO                  http.Handler
+	Login                http.Handler
+	Logout               http.Handler
+	ForgotPassword       http.Handler
+	ResetPassword        http.Handler
+	CreateUserWithInvite http.Handler
+	VerifyInvite         http.Handler
+	InitiateSSO          http.Handler
+	CallbackSSO          http.Handler
+	SettingsSSO          http.Handler
 }
 
 func makeKitHandlers(e FleetEndpoints, opts []kithttp.ServerOption) *fleetHandlers {
@@ -82,16 +76,15 @@ func makeKitHandlers(e FleetEndpoints, opts []kithttp.ServerOption) *fleetHandle
 		return kithttp.NewServer(e, decodeFn, encodeResponse, opts...)
 	}
 	return &fleetHandlers{
-		Login:                        newServer(e.Login, decodeLoginRequest),
-		Logout:                       newServer(e.Logout, decodeNoParamsRequest),
-		ForgotPassword:               newServer(e.ForgotPassword, decodeForgotPasswordRequest),
-		ResetPassword:                newServer(e.ResetPassword, decodeResetPasswordRequest),
-		CreateUserWithInvite:         newServer(e.CreateUserWithInvite, decodeCreateUserRequest),
-		PerformRequiredPasswordReset: newServer(e.PerformRequiredPasswordReset, decodePerformRequiredPasswordResetRequest),
-		VerifyInvite:                 newServer(e.VerifyInvite, decodeVerifyInviteRequest),
-		InitiateSSO:                  newServer(e.InitiateSSO, decodeInitiateSSORequest),
-		CallbackSSO:                  newServer(e.CallbackSSO, decodeCallbackSSORequest),
-		SettingsSSO:                  newServer(e.SSOSettings, decodeNoParamsRequest),
+		Login:                newServer(e.Login, decodeLoginRequest),
+		Logout:               newServer(e.Logout, decodeNoParamsRequest),
+		ForgotPassword:       newServer(e.ForgotPassword, decodeForgotPasswordRequest),
+		ResetPassword:        newServer(e.ResetPassword, decodeResetPasswordRequest),
+		CreateUserWithInvite: newServer(e.CreateUserWithInvite, decodeCreateUserRequest),
+		VerifyInvite:         newServer(e.VerifyInvite, decodeVerifyInviteRequest),
+		InitiateSSO:          newServer(e.InitiateSSO, decodeInitiateSSORequest),
+		CallbackSSO:          newServer(e.CallbackSSO, decodeCallbackSSORequest),
+		SettingsSSO:          newServer(e.SSOSettings, decodeNoParamsRequest),
 	}
 }
 
@@ -269,7 +262,6 @@ func attachFleetAPIRoutes(r *mux.Router, h *fleetHandlers) {
 	r.Handle("/api/v1/fleet/logout", h.Logout).Methods("POST").Name("logout")
 	r.Handle("/api/v1/fleet/forgot_password", h.ForgotPassword).Methods("POST").Name("forgot_password")
 	r.Handle("/api/v1/fleet/reset_password", h.ResetPassword).Methods("POST").Name("reset_password")
-	r.Handle("/api/v1/fleet/perform_required_password_reset", h.PerformRequiredPasswordReset).Methods("POST").Name("perform_required_password_reset")
 	r.Handle("/api/v1/fleet/sso", h.InitiateSSO).Methods("POST").Name("intiate_sso")
 	r.Handle("/api/v1/fleet/sso", h.SettingsSSO).Methods("GET").Name("sso_config")
 	r.Handle("/api/v1/fleet/sso/callback", h.CallbackSSO).Methods("POST").Name("callback_sso")
@@ -437,6 +429,8 @@ func attachNewStyleFleetAPIRoutes(r *mux.Router, svc fleet.Service, logger kitlo
 	// For some reason osquery does not provide a node key with the block data.
 	// Instead the carve session ID should be verified in the service method.
 	ne.POST("/api/_version_/osquery/carve/block", carveBlockEndpoint, carveBlockRequest{})
+
+	ne.POST("/api/_version_/fleet/perform_required_password_reset", performRequiredPasswordResetEndpoint, performRequiredPasswordResetRequest{})
 }
 
 // TODO: this duplicates the one in makeKitHandler
