+++ server/service/transport_users.go
@@ -4,8 +4,6 @@ import (
 	"context"
 	"encoding/json"
 	"net/http"
-
-	"github.com/fleetdm/fleet/v4/server/contexts/ctxerr"
 )
 
 func decodeCreateUserRequest(ctx context.Context, r *http.Request) (interface{}, error) {
@@ -14,14 +12,6 @@ func decodeCreateUserRequest(ctx context.Context, r *http.Request) (interface{},
 		return nil, err
 	}
 
-	return req, nil
-}
-
-func decodePerformRequiredPasswordResetRequest(ctx context.Context, r *http.Request) (interface{}, error) {
-	var req performRequiredPasswordResetRequest
-	if err := json.NewDecoder(r.Body).Decode(&req); err != nil {
-		return nil, ctxerr.Wrap(ctx, err, "decoding JSON")
-	}
 	return req, nil
 }
